src/gh-issue-importer/requirements.txt
requests
pandas 
python-dotenv
